+++ src/modules/legal/legal.module.ts
@@ -1,6 +1,7 @@
-﻿import { Module } from '@nestjs/common';
+﻿import { forwardRef, Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
 
+import { AuthModule } from '../auth/auth.module';
 import { ILegalTermsRepositoryToken } from '../../domain/legal/interfaces/legal-terms.repository.interface';
 import { LegalTermsRepository } from '../../infrastructure/legal/legal-terms.repository';
 import { LegalTermEntity } from '../../infrastructure/legal/entities/legal-term.entity';
@@ -8,7 +9,7 @@ import { LegalTermsService } from './legal-terms.service';
 import { LegalTermsController } from './api/controllers/legal-terms.controller';
 
 @Module({
-  imports: [TypeOrmModule.forFeature([LegalTermEntity])],
+  imports: [TypeOrmModule.forFeature([LegalTermEntity]), forwardRef(() => AuthModule)],
   controllers: [LegalTermsController],
   providers: [
     LegalTermsService,